Fix incompatibility when used with compileSdk 33 projects
The layoutlib version which Paparazzi is currently using, only supports
up to SDK level 32. When a project using compileSdkVersion 33+ tries to
use Paparazzi, lots of things break.

This is caused by Paparazzi Gradle plugin passing through the
compileSdkVersion directly through to layoutlib. This PR changes
that behavior, by coercing the SDK_INT (and CODENAME) values to be
API 32 at most (to match layoutlib's support).

This means that projects using compileSdkVersion 33 will remain to work,
just running on an API 32 host.

# Conflicts:
#	gradle/libs.versions.toml
#	paparazzi/paparazzi/src/main/java/app/cash/paparazzi/Paparazzi.kt
#	sample/build.gradle